[Adjustment][Merv] Move protected route to routes
